refactor(LC033): split frozen-set analyzer and fixer helpers
- separate analyzer field and usage analysis
- separate fixer application and rewrite logic
- keep analyzer/fixer behavior unchanged while reducing local complexity
